backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
aiogram==3.3.0
aiohttp==3.9.1
httpx==0.25.1
# ton-blockchain==0.0.1
# tonpy==0.2.0
# pytonlib==0.1.0
cryptography==41.0.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
cors==1.0.1
